fix(rtypes): serve configured A-record TTL immediately after add
Add stored the RRset TTL as uint32 in the record map while Lookup only
type-asserts float64, so freshly added A records were served with the
default TTL 3600 until a persistence round-trip converted the value to
float64 (i.e. until restart). Store the TTL as float64 at write time so the
in-memory shape matches the persisted shape, and add a red/green regression
test that pins the served TTL right after add.

+++ zone/rtypes/a.go
@@ -93,7 +93,7 @@ func (ARecord) Add(zone, name string, value interface{}, ttl *uint32) error {
 	for _, r := range currentList {
 		listToStore = append(listToStore, map[string]interface{}{
 			"ip":  r.IP,
-			"ttl": r.TTL,
+			"ttl": float64(r.TTL),
 		})
 	}
